+++ readium/lcp/r2-lcp/src/main/java/org/readium/r2/lcp/service/LicensesRepository.kt
@@ -10,11 +10,36 @@
 package org.readium.r2.lcp.service
 
 import org.readium.r2.lcp.license.model.LicenseDocument
+import org.readium.r2.lcp.persistence.LcpDao
+import org.readium.r2.lcp.persistence.License
 
-internal interface LicensesRepository {
-    fun addLicense(license: LicenseDocument)
-    fun copiesLeft(licenseId: String) : Int?
-    fun setCopiesLeft(quantity: Int, licenseId: String)
-    fun printsLeft(licenseId: String) : Int?
-    fun setPrintsLeft(quantity: Int, licenseId: String)
+internal class LicensesRepository(private val lcpDao: LcpDao) {
+
+    suspend fun addLicense(licenseDocument: LicenseDocument) {
+        if (lcpDao.exists(licenseDocument.id) != null) {
+            return
+        }
+        val license = License(
+            licenseId = licenseDocument.id,
+            rightPrint = licenseDocument.rights.print,
+            rightCopy = licenseDocument.rights.copy
+        )
+        lcpDao.addLicense(license)
+    }
+
+    fun copiesLeft(licenseId: String): Int? {
+        return lcpDao.getCopiesLeft(licenseId)
+    }
+
+    fun setCopiesLeft(quantity: Int, licenseId: String) {
+        lcpDao.setCopiesLeft(quantity, licenseId)
+    }
+
+    fun printsLeft(licenseId: String) : Int? {
+        return lcpDao.getPrintsLeft(licenseId)
+    }
+
+    fun setPrintsLeft(quantity: Int, licenseId: String) {
+        lcpDao.setPrintsLeft(quantity, licenseId)
+    }
 }