+++ readium/lcp/r2-lcp/src/main/java/org/readium/r2/lcp/service/PassphrasesRepository.kt
@@ -9,9 +9,35 @@
 
 package org.readium.r2.lcp.service
 
-internal interface PassphrasesRepository {
-    fun passphrase(licenseId: String) : String?
-    fun passphrases(userId: String) : List<String>
-    fun allPassphrases(): List<String>
-    fun addPassphrase(passphraseHash: String, licenseId: String?, provider: String?, userId: String?)
+import org.readium.r2.lcp.persistence.LcpDao
+import org.readium.r2.lcp.persistence.Passphrase
+
+internal class PassphrasesRepository(private val lcpDao: LcpDao) {
+
+    suspend fun passphrase(licenseId: String): String? {
+        return lcpDao.passphrase(licenseId)
+    }
+
+    suspend fun passphrases(userId: String): List<String> {
+        return lcpDao.passphrases(userId)
+    }
+
+    suspend fun allPassphrases(): List<String> {
+        return lcpDao.allPassphrases()
+    }
+
+    suspend fun addPassphrase(
+        passphraseHash: String,
+        licenseId: String,
+        provider: String,
+        userId: String?
+    ) {
+        val transaction = Passphrase(
+            licenseId = licenseId,
+            provider = provider,
+            userId = userId,
+            passphrase = passphraseHash
+        )
+        lcpDao.addPassphrase(transaction)
+    }
 }